Swap to centralized model, cache change
After some consideration decided to move the peerlist code to my own
server instead of the site owners, will make a solution for using both.
Fixed the deletion of disconnected peers in the mean time.

Changed the cache from "localStorage" to in memory javascript, the
library no works best only with async pages.